Show accounts under sub ledger detail view Show sub ledgers in separate view under ledgers Removed unused resolver Fix issue not displaying conflict error when creating sub ledger
