--- src/app/accounting/accounts/account.resolver.ts
@@ -1,31 +1,0 @@
-/**
- * Copyright 2017 The Mifos Initiative.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {Injectable} from '@angular/core';
-import {Resolve, RouterStateSnapshot, ActivatedRouteSnapshot} from '@angular/router';
-import {Account} from '../../../services/accounting/domain/account.model';
-import {Observable} from 'rxjs';
-import {AccountingService} from '../../../services/accounting/accounting.service';
-
-@Injectable()
-export class AccountResolver implements Resolve<Account>{
-
-  constructor(private accountingService: AccountingService) {}
-
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Account> {
-    return this.accountingService.findAccount(route.params['id']);
-  }
-}
