+++ src/app/accounting/general-ledger.component.html
@@ -45,7 +45,7 @@
           <div flex-gt-md="70">
             <div layout="row">
               <div layout="column" flex-gt-md="100">
-                <h3 translate>Subledger</h3>
+                <h3 translate>Ledger</h3>
                 <fims-data-table flex (onActionCellClick)="rowSelect($event)" [columns]="columns" [data]="ledgerData | async" sortBy="identifier"></fims-data-table>
               </div>
             </div>
