+++ src/app/accounting/store/account/effects/route.effects.ts
@@ -28,7 +28,13 @@ export class AccountRouteEffects {
 
   @Effect({ dispatch: false })
   createAccountSuccess$: Observable<Action> = this.actions$
-    .ofType(accountActions.CREATE_SUCCESS, accountActions.UPDATE_SUCCESS)
+    .ofType(accountActions.CREATE_SUCCESS)
+    .map(action => action.payload)
+    .do(payload => this.router.navigate(['../../'], { relativeTo: payload.activatedRoute }));
+
+  @Effect({ dispatch: false })
+  updateAccountSuccess$: Observable<Action> = this.actions$
+    .ofType(accountActions.UPDATE_SUCCESS)
     .map(action => action.payload)
     .do(payload => this.router.navigate(['../'], { relativeTo: payload.activatedRoute }));
 
@@ -36,5 +42,5 @@ export class AccountRouteEffects {
   deleteAccountSuccess$: Observable<Action> = this.actions$
     .ofType(accountActions.DELETE_SUCCESS)
     .map(action => action.payload)
-    .do(payload => this.router.navigate(['../../../ledgers/detail', payload.resource.ledger, 'accounts'], { relativeTo: payload.activatedRoute }));
+    .do(payload => this.router.navigate(['../../../ledgers/detail', payload.resource.ledger], { relativeTo: payload.activatedRoute }));
 }
