+++ src/assets/i18n/es.json
@@ -8,6 +8,7 @@
   "Account is going to be saved": "",
   "Account settings": "Configuración de cuentas",
   "Account": "Cuenta",
+  "Account with value": "Account {{value}}",
   "Accounting": "Contabilidad",
   "Accounts for ledger": "Cuentas para el libro principal",
   "Accounts": "Cuentas",
@@ -151,6 +152,7 @@
   "DISBURSE LOAN": "DESEMBOLSAR PRÉSTAMO",
   "Disbursement fee": "",
   "Do you want to delete ledger {{value}}": "¿Quiere eliminar el libro {{value}} contable?",
+  "Do you want to delete this ledger?": "",
   "Do you want to delete this account?": "",
   "Edit account": "Editar cuenta",
   "Edit charge": "Editar cargo",
@@ -287,7 +289,7 @@
   "Must have decimal places": "",
   "Name": "Nombre",
   "Navigate to ledger of this account": "Navegar a través del libro principal de ésta cuenta",
-  "Navigate to reference account": "Navegar en la cuenta de referencia\n",
+  "Navigate to reference account": "Navegar en la cuenta de referencia",
   "Navigation": "Navegación",
   "New Password": "Nueva contraseña",
   "Next repayment": "Próximo pagos",
@@ -447,6 +449,7 @@
   "View payments": "Ver pagos",
   "View roles": "Ver roles",
   "View trial balance": "Ver balance de comprobación",
+  "View subledgers": "",
   "We are very sorry, it seems there is a problem with our servers. Please contact your administrator if the problem occurs.": "Lo sentimos, pero hay un problema con nuestros servidores. Por favor contacte a su administrador si el problema vuelve a suceder.",
   "Wednesday": "Miércoles",
   "WEEKS": "SEMANAS",
